# Add NullableAnnotation/WithNullableAnnotation APIs to ITypeSymbol.