--- a/src/Compilers/CSharp/Portable/CSharpExtensions.cs
+++ b/src/Compilers/CSharp/Portable/CSharpExtensions.cs
@@ -1152,7 +1152,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             var csmodel = semanticModel as CSharpSemanticModel;
             if (csmodel != null)
             {
-                return csmodel.ClassifyConversion(expression, (TypeSymbol)destination, isExplicitInSource);
+                return csmodel.ClassifyConversion(expression, destination, isExplicitInSource);
             }
             else
             {
@@ -1170,7 +1170,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             var csmodel = semanticModel as CSharpSemanticModel;
             if (csmodel != null)
             {
-                return csmodel.ClassifyConversion(position, expression, (TypeSymbol)destination, isExplicitInSource);
+                return csmodel.ClassifyConversion(position, expression, destination, isExplicitInSource);
             }
             else
             {