--- a/src/Compilers/CSharp/Portable/Compilation/AwaitExpressionInfo.cs
+++ b/src/Compilers/CSharp/Portable/Compilation/AwaitExpressionInfo.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
+using Microsoft.CodeAnalysis.CSharp.Symbols;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp