--- a/src/Compilers/CSharp/Portable/Compilation/TypeInfo.cs
+++ b/src/Compilers/CSharp/Portable/Compilation/TypeInfo.cs
@@ -48,7 +48,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
         public static implicit operator TypeInfo(CSharpTypeInfo info)
         {
-            return new TypeInfo(info.Type, info.ConvertedType, info.Nullability, info.ConvertedNullability);
+            return new TypeInfo(info.Type.GetPublicSymbol(), info.ConvertedType.GetPublicSymbol(), info.Nullability, info.ConvertedNullability);
         }
 
         public override bool Equals(object obj)