--- a/src/Compilers/CSharp/Test/Semantic/Semantics/OperatorTests.cs
+++ b/src/Compilers/CSharp/Test/Semantic/Semantics/OperatorTests.cs
@@ -7245,7 +7245,6 @@ public class RubyTime
 class Module1
 {
     void Test(int x)
-
     {
         var z1 = -x;
         var z2 = --x;