… src/Compilers/CSharp/Portable/Compilation/CSharpCompilation.cs
… src/Compilers/CSharp/Portable/Compilation/CSharpSemanticModel.cs
… src/Compilers/CSharp/Portable/Compiler/ClsComplianceChecker.cs
… src/Compilers/CSharp/Portable/Emitter/EditAndContinue/CSharpDefinitionMap.cs
… src/Compilers/CSharp/Portable/Emitter/EditAndContinue/CSharpSymbolChanges.cs
… src/Compilers/CSharp/Portable/Emitter/EditAndContinue/CSharpSymbolMatcher.cs
… src/Compilers/CSharp/Portable/Emitter/EditAndContinue/PEDeltaAssemblyBuilder.cs
… src/Compilers/CSharp/Portable/Emitter/Model/AttributeDataAdapter.cs
… src/Compilers/CSharp/Portable/Emitter/Model/CustomModifierAdapter.cs
… src/Compilers/CSharp/Portable/Emitter/Model/ExpandedVarargsMethodReference.cs
… src/Compilers/CSharp/Portable/Emitter/Model/NamedTypeSymbolAdapter.cs
… src/Compilers/CSharp/Portable/Emitter/Model/PEAssemblyBuilder.cs
… src/Compilers/CSharp/Portable/Emitter/Model/PEModuleBuilder.cs
… src/Compilers/CSharp/Portable/Emitter/Model/PENetModuleBuilder.cs
… src/Compilers/CSharp/Portable/Emitter/NoPia/EmbeddedEvent.cs
… src/Compilers/CSharp/Portable/FlowAnalysis/CSharpDataFlowAnalysis.cs
… src/Compilers/CSharp/Portable/FlowAnalysis/DataFlowsOutWalker.cs
… src/Compilers/CSharp/Portable/FlowAnalysis/EmptyStructTypeCache.cs
… src/Compilers/CSharp/Portable/FlowAnalysis/NullableWalker.SnapshotManager.cs
… src/Compilers/CSharp/Portable/FlowAnalysis/NullableWalker.cs
… src/Compilers/CSharp/Portable/Generated/BoundNodes.xml.Generated.cs
… src/Compilers/CSharp/Portable/Lowering/AsyncRewriter/AsyncConstructor.cs
… src/Compilers/CSharp/Portable/Lowering/AsyncRewriter/AsyncMethodToStateMachineRewriter.cs
… src/Compilers/CSharp/Portable/Lowering/IteratorRewriter/IteratorConstructor.cs
… src/Compilers/CSharp/Portable/Lowering/IteratorRewriter/IteratorFinallyMethodSymbol.cs
… src/Compilers/CSharp/Portable/Lowering/LambdaRewriter/SynthesizedClosureEnvironment.cs
… src/Compilers/CSharp/Portable/Lowering/LambdaRewriter/SynthesizedClosureEnvironmentConstructor.cs
… src/Compilers/CSharp/Portable/Lowering/LambdaRewriter/SynthesizedClosureMethod.cs
… src/Compilers/CSharp/Portable/Lowering/LocalRewriter/DynamicSiteContainer.cs
… src/Compilers/CSharp/Portable/Lowering/LocalRewriter/LocalRewriter_Call.cs