This change also separates implementation of ISymbol interfaces from internal symbols in C# compiler.

Perf results from this change https://github.com/dotnet/roslyn/pull/39498#issuecomment-546450046.