@333fred 2019-10-25
>         private ImmutableArray<Symbol> LookupSymbolsInternal(

Do we want to consider making this return `ImmutableArray<ISymbol>` instead of `Symbol`, so that we don't need to allocate 2 arrays, one for symbols and one for ISymbols?

---
Refers to: src/Compilers/CSharp/Portable/Compilation/CSharpSemanticModel.cs:1460 in c2f38f5. [](commit_id = c2f38f5744798afe426f7f3b9ed7a4453f19a671, deletion_comment = False)