@agocke 2019-10-25
Perf results from this change:

Before
```
|                      Method | UnrollFactor |        Mean |     Error |    StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 |  Allocated |
|---------------------------- |------------- |------------:|----------:|----------:|------------:|------------:|------------:|------------:|-----------:|----------:|-----------:|
|       CompileMethodsAndEmit |           16 |  5,740.7 ms |  61.96 ms |  92.75 ms |  5,704.7 ms |  5,635.5 ms |  5,980.8 ms |  65000.0000 | 18000.0000 |         - |  505.62 MB |
|           SerializeMetadata |           16 |    356.2 ms |   8.64 ms |  25.48 ms |    356.3 ms |    316.1 ms |    413.3 ms |   2000.0000 |  1000.0000 |         - |   33.21 MB |
|              GetDiagnostics |            1 |  4,886.4 ms |  48.35 ms | 123.94 ms |  4,896.2 ms |  4,611.7 ms |  5,223.6 ms |  51000.0000 | 14000.0000 |         - |  389.34 MB |
| GetDiagnosticsWithAnalyzers |            1 | 16,766.2 ms | 164.08 ms | 283.02 ms | 16,789.6 ms | 16,164.3 ms | 17,240.5 ms | 184000.0000 | 50000.0000 | 2000.0000 | 1395.41 MB |
```

After:
```
|                      Method | UnrollFactor |        Mean |     Error |    StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|---------------------------- |------------- |------------:|----------:|----------:|------------:|------------:|------------:|------------:|-----------:|----------:|----------:|
|       CompileMethodsAndEmit |           16 |  6,388.7 ms |  63.21 ms | 156.23 ms |  6,384.2 ms |  6,032.2 ms |  6,691.1 ms |  65000.0000 | 18000.0000 |         - | 507.29 MB |
|           SerializeMetadata |           16 |    358.6 ms |   7.36 ms |  21.71 ms |    361.1 ms |    312.7 ms |    411.1 ms |   2000.0000 |  1000.0000 |         - |  33.25 MB |
|              GetDiagnostics |            1 |  4,775.7 ms |  47.56 ms | 106.37 ms |  4,773.2 ms |  4,588.0 ms |  5,072.2 ms |  52000.0000 | 13000.0000 |         - | 390.19 MB |
| GetDiagnosticsWithAnalyzers |            1 | 16,223.9 ms | 159.99 ms | 370.79 ms | 16,100.5 ms | 15,699.6 ms | 17,135.8 ms | 187000.0000 | 57000.0000 | 2000.0000 |   1413 MB |
```

My conclusion: no significant change in overall performance. A small (1.5%) increase in allocations when compiling with analyzers. I would say that's acceptable. I think we can get allocation wins from other areas in the pipeline that more than make up for it.