@gafter 2019-10-28
>             var c = CreateCompilation("");

Left off here in Iteration 2.  Will pick up later.

---
Refers to: src/Compilers/CSharp/Test/Symbol/Compilation/QueryClauseInfoTests.cs:16 in cacb2fd. [](commit_id = cacb2fdb4bb68f808f19a0a74cf87e06e5e7dbca, deletion_comment = True)