@gafter 2019-10-29
>             var comp = (Compilation)CreateCompilation(source);

Places where you changed the code to
``` c#
var comp = (Compilation)CreateCompilation(...);
```
feels like it might have been better
``` c#
Compilation comp = CreateCompilation(...);
```
Because then it is clear that this is an implicit conversion.

---
Refers to: src/Compilers/CSharp/Test/Symbol/Compilation/SemanticModelAPITests.cs:1196 in cacb2fd. [](commit_id = cacb2fdb4bb68f808f19a0a74cf87e06e5e7dbca, deletion_comment = False)