@tmat 2019-10-30
Reviewed iteration 6.