@AlekseyTs 2019-10-30
@jasonmalinowski, @ryzngard Please provide feedback at least on the new public API on ITypeSymbol and on changes on IDE side. Thanks.