@CyrusNajmabadi 2019-10-30
@jasonmalinowski do you intend to make a followup PR that moves the IDE onto this api first?  i.e. are you going to prove this out before this gets checked in?  or will this be checked in, and then you'll adopt at your own pace?  Thanks!